Display Saved Views Widget within a viewport
