+++ packages/test-app-frontend/src/App/common/Overlap.css
@@ -1,0 +1,13 @@
+/*---------------------------------------------------------------------------------------------
+* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+* See LICENSE.md in the project root for license terms and full copyright notice.
+*--------------------------------------------------------------------------------------------*/
+
+.overlap {
+  display: grid;
+}
+
+.overlap > * {
+  grid-row: 1 / 1;
+  grid-column: 1 / 1;
+}
